requirements.txt
graphics.py
nose
sphinx
setuptools~=50.3.2